requirements.txt
streamlit>=1.28.0
pillow>=9.0.0
pandas>=1.5.0
numpy>=1.24.0
plotly>=5.15.0
matplotlib>=3.7.0
scipy>=1.10.0
spacy>=3.6.0
scikit-learn>=1.3.0
requests>=2.31.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0